Add functions to implement: 1. Normalize the market trend data. 2. Detect sales opportunity function. 3. Once an opportunity is detected, simple message is display on webpage.
At this moment, the code still has issue if no data at the begining of
the time period.
For example, the code does not work well for the following category
selection.
member_service/sales_opportunities/?cat1=7&cat2=1032228&cat3=122374007

I had to add code to use middle year data to do normalization,
somehow, it does not work.

Changes to be committed:
	modified:   member_service/views.py
	modified:   templates/member_service/sales_opportunities.html